GraphView refactoring Spited into components files Removed unnecessary dependencies
