+++ app/index.html
@@ -12,10 +12,6 @@
 
     <script src="lib/d3/d3.js"></script>
 
-    <script src="lib/sigma.js/sigma.min.js"></script>
-    <script src="lib/sigma.js/plugins/sigma.renderers.customShapes.min.js"></script>
-    <script src="lib/sigma.js/plugins/sigma.layout.forceAtlas2.min.js"></script>
-
     <script src="scripts/app.js"></script>
 </head>
 <body>
