+++ app/scripts/components/Common/Button.js
@@ -1,0 +1,9 @@
+
+var Button = React.createClass({
+    render() {
+        var {title, ...props} = this.props;
+        return <input type="button" {...props} value={title} />;
+    }
+});
+
+export default Button;
