+++ app/scripts/components/Common/Common.js
@@ -1,0 +1,7 @@
+import Button from "./Button";
+import Field from "./Field";
+
+export default {
+    Button,
+    Field
+};
