+++ app/scripts/components/Common/Field.js
@@ -1,0 +1,7 @@
+var Field = React.createClass({
+    render() {
+        return <input {...this.props} />;
+    }
+});
+
+export default Field;
