+++ app/scripts/components/GraphView/GraphView.js
@@ -1,6 +1,6 @@
 import Graph from "../D3Graph";
 import Identifier from "../../model/Identifier";
-
+import SearchMenu from "./SearchMenu";
 
 var GraphView = React.createClass({
 
@@ -132,194 +132,6 @@ var GraphView = React.createClass({
         this.graph = null;
     }
 
-});
-
-/**
- * kvValue = {String: [String]}
- * this.state.max_depth = Number
- * this.state.results_filter = [String]
- * this.state.match_links = [{key: [values]}, ]
- * this.state.match_metadata = [{key: [values]}, ]
- * this.state.match_terminal = [{key: [values]}, ]
- */
-
-const fieldPairs = [{
-    name: "Match Identifiers Metadata",
-    paramsKey: "match_metadata"
-}, {
-    paramsKey: "match_links",
-    name: "Match Links Metadata"
-}, {
-    paramsKey: "match_terminal",
-    name: "Match Identifiers Metadata"
-}];
-
-
-var SearchMenu = React.createClass({
-
-    getInitialState() {
-        return {
-            params: {
-                max_depth: 1,
-                results_filter: [],
-                match_metadata: [],
-                match_links: [],
-                match_terminal: []
-            }
-        }
-    },
-
-    paramsChanged(paramsKey, value) {
-        var params = this.state.params;
-
-        params[paramsKey] = value;
-
-        this.setState({params});
-    },
-
-    _addField(key) {
-        var params = this.state.params;
-
-        params[key].push({key: "", value: []});
-
-        this.setState({params})
-    },
-
-    _removeField(key, index) {
-        var fields = this.state.params[key];
-        fields.splice(index, 1);
-        this.setState({fields});
-    },
-
-    _normaliseValues(params) {
-        var {max_depth, results_filter, ...toNormalise} = params;
-
-        Object.keys(toNormalise).map((key) => {
-            toNormalise[key] = toNormalise[key].reduce((obj, val) => {
-                obj[val.key] = val.value;
-            }, {});
-        });
-
-        return {max_depth, results_filter, ...toNormalise};
-    },
-
-    _onSubmit(event) {
-        event.preventDefault();
-        console.log(this.state.params);
-        this.props.onSubmit(this._normaliseValues(this.state.params));
-    },
-
-    render() {
-        var {x: left=50000, y: top=50000} = this.props.position || {};
-
-        var fields = fieldPairs.map(({paramsKey, name}) => {
-            return (
-                <AddableFieldPair
-                    fields={this.state.params[paramsKey]}
-                    addField={() => this._addField(paramsKey)}
-                    removeField={(index) => this._removeField(paramsKey, index)}
-                    onChange={(fields) => this.paramsChanged(paramsKey, fields)}
-                    name={name}
-                />
-            )
-        });
-
-        return (
-            <div className="card menu" style={{position: "fixed", top, left}}>
-                <form onSubmit={this._onSubmit} onChange={(fields) => console.log(fields)}>
-                    <div>Search for Identifiers</div>
-                    <div>
-                        <p>
-                            <span>Max Depth </span>
-                            <Field
-                                type="number"
-                                onChange={({target: {value}}) => this.paramsChanged("max_depth", parseInt(value, 10))}
-                                value={this.state.params.max_depth}
-                            />
-                        </p>
-                    </div>
-                    <div>
-                        <p>
-                            <span>Filter Metadata</span>
-                            <Field
-                                onChange={({target: {value}}) => this.paramsChanged("results_filter", value.split(" "))}
-                                value={this.state.params.results_filter.join(" ")}
-                            />
-                        </p>
-                    </div>
-                    {fields}
-                    <Button type="submit" onClick={this._onSubmit} title="Search" />
-                </form>
-            </div>
-        )
-    }
-});
-
-var AddableFieldPair = React.createClass({
-
-    _addField() {
-        this.props.addField();
-    },
-
-    _removeField(index) {
-        this.props.removeField(index);
-    },
-
-    _keyChanged(index, key) {
-        var fields = this.props.fields,
-            {value} = fields[index];
-
-        fields[index] = {key, value};
-        this.props.onChange(fields);
-    },
-
-    _valueChanged(index, value) {
-        var fields = this.props.fields,
-            {key} = fields[index];
-
-        fields[index] = {key, value:value.split(" ")};
-        this.props.onChange(fields);
-    },
-
-    render() {
-        var fields = this.props.fields.map((field, index) => {
-            var {key, value} = field;
-
-            return (
-                    <div key={index}>
-                        <Field onChange={({target: {value}}) => this._keyChanged(index, value)} value={key} />
-                        <Field onChange={({target: {value}}) => this._valueChanged(index, value)} value={value.join(" ")} />
-                        <Button onClick={() => this._removeField(index)} title="-" />
-                    </div>
-                )
-        });
-        return (
-            <div>
-                <div>
-                    <p>
-                        <span>{this.props.name}</span>
-                        <Button onClick={this._addField} title="+" />
-                    </p>
-                </div>
-                <div>
-                    {fields}
-                </div>
-            </div>
-        )
-    }
-});
-
-var Field = React.createClass({
-    render() {
-        return <input {...this.props} />;
-    }
-});
-
-var Button = React.createClass({
-    render() {
-        var {title, ...props} = this.props;
-        return <input type="button" {...props} value={title} />;
-    }
 });
